+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntitySearchField.java
@@ -1167,12 +1167,12 @@ public final class EntitySearchField extends HintTextField {
 		}
 
 		private void showProgressBarDelayed() {
-			showProgressBarAction = delayedAction(searchRefreshProgressBarDelay, () -> {
+			showProgressBarAction = delayedAction(() -> {
 				setLayout(new BorderLayout());
 				add(progressBar, BorderLayout.CENTER);
 				revalidate();
 				repaint();
-			});
+			}, searchRefreshProgressBarDelay);
 		}
 
 		private void hideProgressBar() {